pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "torcpy"
version = "0.1.0"
description = "Distributed workflow orchestration system"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "aiosqlite>=0.20",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "pydantic>=2.10",
    "httpx>=0.28",
    "rich-click>=1.8",
    "rich>=13.0",
    "pyyaml>=6.0",
    "json5>=0.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]
docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.27",
    "mkdocs-mermaid2-plugin>=1.1",
]

[project.scripts]
torcpy = "torcpy.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/torcpy"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
